reorganized api calls
