+++ app/api/client/auth.ts
