+++ app/auth/sign-in.tsx
@@ -5,7 +5,7 @@ import { Button, TextField, Grid, Typography, Paper, Box } from "@mui/material";
 import { useForm } from "react-hook-form";
 import { yupResolver } from "@hookform/resolvers/yup";
 import { emailSchema } from "app/utils/schemas";
-import { sendPasswordlessLoginEmail, signInWithLink } from "app/api/auth";
+import { sendPasswordlessLoginEmail, signInWithLink } from "app/api/client/auth";
 import { useRecoilState } from "recoil";
 import { UserState } from "app/recoil-store";
