+++ app/client-hooks/user-api.ts
@@ -1,6 +1,6 @@
 "use client";
 
-import { privateUserCollection } from "app/api/config";
+import { privateUserCollection } from "app/api/server/config";
 import { PrivateUserData } from "app/types";
 import { useState, useCallback } from "react";
